ai-data-validator/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
openai==1.7.2
pinecone-client==2.2.4
langchain==0.1.0
pandas==2.1.4
polars==0.20.2
PyMuPDF==1.23.14
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10
aiofiles==23.2.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
tiktoken==0.5.2
requests==2.31.0
numpy==1.26.2
scikit-learn==1.3.2
nltk==3.8.1
textstat==0.7.3
regex==2023.10.3
